+++ testData/org/elixir_lang/reference/module/cyclic_alias/mutual_use.ex
@@ -1,0 +1,8 @@
+defmodule MutualUse do
+  use CycleA
+  use CycleB
+
+  Miss<caret>ing
+
+  @a 1
+end