+++ testData/org/elixir_lang/reference/module/cyclic_alias/self_referential.ex
@@ -1,0 +1,7 @@
+defmodule Cyclic.Usage do
+  alias Missing
+
+  Miss<caret>ing
+
+  @a 1
+end