+++ testData/org/elixir_lang/reference/module/cyclic_alias/transitive.ex
@@ -1,0 +1,8 @@
+defmodule Transitive.Usage do
+  alias MyNamespace.Referenced
+  alias Referenced, as: Refd
+
+  Re<caret>fd
+
+  @a 1
+end